Extracted actions and events to separate file
